+++ autokeras/image_supervised.py
